rcnn/requirements.txt
opencv-contrib-python
keras
tensorflow-gpu
numpy
pandas
matplotlib
